chrome/test/data/android/webvr_instrumentation/html/test_screen_taps_registered.html

<!doctype html>
<!--
Tests that either screen taps are registered in VR when viewer is Cardboard
or that controller clicks are registered as screen taps in VR when viewer is
Daydream View.
-->
<html>
  <head>
    <link rel="stylesheet" type="text/css" href="../resources/webvr_e2e.css">
  </head>
  <body>
    <canvas id="webgl-canvas"></canvas>
    <script src="../../../../../../third_party/WebKit/LayoutTests/resources/testharness.js"></script>
    <script src="../resources/webvr_e2e.js"></script>
    <script src="../resources/webvr_boilerplate.js"></script>
    <script>
      var t = async_test("Screen taps/clicks registered when in VR");
      window.addEventListener("vrdisplaypresentchange",
          () => {finishJavaScriptStep();}, false);
      var numTaps = 0;
      webglCanvas.addEventListener("click",
          () => {
            numTaps++;
            // Notify Java when we've received the tap after entering VR
            if (numTaps == 2) {
              t.done();
            }
          }, false);

      function stepVerifyNoInitialTaps() {
        t.step( () => {
          assert_equals(numTaps, 0, "No initial taps");
        });
        finishJavaScriptStep();
      }
    </script>
  </body>
</html>
